+++ src/hooks.spec.tsx
@@ -15,16 +15,32 @@
  */
 
 /// <reference types="jest" />
-
 import * as React from 'react';
 import { act } from 'react-dom/test-utils';
 import { render, renderHook, screen, waitFor } from '@testing-library/react';
 import '@testing-library/jest-dom';
 
 import { OptimizelyProvider } from './Provider';
-import { OnReadyResult, ReactSDKClient, VariableValuesObject } from './client';
-import { useExperiment, useFeature, useDecision, useTrackEvent, hooksLogger } from './hooks';
+import { NotReadyReason, OnReadyResult, ReactSDKClient, VariableValuesObject } from './client';
+import { useExperiment, useFeature, useDecision, useTrackEvent } from './hooks';
 import { OptimizelyDecision } from './utils';
+import { getLogger } from '@optimizely/optimizely-sdk';
+
+jest.mock('@optimizely/optimizely-sdk', () => {
+  const originalModule = jest.requireActual('@optimizely/optimizely-sdk');
+  return {
+    ...originalModule,
+    getLogger: jest.fn().mockReturnValue({
+      error: jest.fn(),
+      warn: jest.fn(),
+      info: jest.fn(),
+      debug: jest.fn(),
+    }),
+  };
+});
+
+const hooksLogger = getLogger('ReactSDK');
+
 const defaultDecision: OptimizelyDecision = {
   enabled: false,
   variables: {},
@@ -71,7 +87,6 @@ describe('hooks', () => {
   let notificationListenerCallbacks: Array<() => void>;
   let optimizelyMock: ReactSDKClient;
   let readySuccess: boolean;
-  // let reason: NotReadyReason;
   let userUpdateCallbacks: Array<() => void>;
   let UseExperimentLoggingComponent: React.FunctionComponent<any>;
   let UseFeatureLoggingComponent: React.FunctionComponent<any>;
@@ -80,25 +95,41 @@ describe('hooks', () => {
   let forcedVariationUpdateCallbacks: Array<() => void>;
   let decideMock: jest.Mock<OptimizelyDecision>;
   let setForcedDecisionMock: jest.Mock<void>;
-  let hooksLoggerErrorSpy: jest.SpyInstance;
   const REJECTION_REASON = 'A rejection reason you should never see in the test runner';
 
   beforeEach(() => {
-    getOnReadyPromise = ({ timeout = 0 }: any): Promise<OnReadyResult> =>
-      new Promise((resolve) => {
-        setTimeout(function () {
-          resolve(
-            Object.assign(
-              {
-                success: readySuccess,
-              },
-              !readySuccess && {
-                dataReadyPromise: new Promise((r) => setTimeout(r, mockDelay)),
-              }
-            )
-          );
-        }, timeout || mockDelay);
+    getOnReadyPromise = ({ timeout = 0 }: any): Promise<OnReadyResult> => {
+      const timeoutPromise = new Promise<OnReadyResult>((resolve) => {
+        setTimeout(
+          () => {
+            resolve({
+              success: false,
+              reason: NotReadyReason.TIMEOUT,
+              dataReadyPromise: new Promise((r) =>
+                setTimeout(
+                  () =>
+                    r({
+                      success: readySuccess,
+                    }),
+                  mockDelay
+                )
+              ),
+            });
+          },
+          timeout || mockDelay + 1
+        );
       });
+
+      const clientAndUserReadyPromise = new Promise<OnReadyResult>((resolve) => {
+        setTimeout(() => {
+          resolve({
+            success: readySuccess,
+          });
+        }, mockDelay);
+      });
+
+      return Promise.race([clientAndUserReadyPromise, timeoutPromise]);
+    };
     activateMock = jest.fn();
     isFeatureEnabledMock = jest.fn();
     featureVariables = mockFeatureVariables;
@@ -109,7 +140,6 @@ describe('hooks', () => {
     forcedVariationUpdateCallbacks = [];
     decideMock = jest.fn();
     setForcedDecisionMock = jest.fn();
-    hooksLoggerErrorSpy = jest.spyOn(hooksLogger, 'error');
     optimizelyMock = {
       activate: activateMock,
       onReady: jest.fn().mockImplementation((config) => getOnReadyPromise(config || {})),
@@ -169,59 +199,54 @@ describe('hooks', () => {
       (res) => res.dataReadyPromise,
       (err) => null
     );
-    hooksLoggerErrorSpy.mockReset();
+    jest.resetAllMocks();
   });
 
   describe('useExperiment', () => {
     it('should return a variation when activate returns a variation', async () => {
       activateMock.mockReturnValue('12345');
 
-      const { container } = render(
+      render(
         <OptimizelyProvider optimizely={optimizelyMock}>
           <MyExperimentComponent />
         </OptimizelyProvider>
       );
-      await optimizelyMock.onReady();
       await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('12345|true|false'));
     });
 
     it('should return null when activate returns null', async () => {
       activateMock.mockReturnValue(null);
       featureVariables = {};
-      const { container } = render(
+      render(
         <OptimizelyProvider optimizely={optimizelyMock}>
           <MyExperimentComponent />
         </OptimizelyProvider>
       );
-      await optimizelyMock.onReady();
       await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('null|true|false'));
     });
 
     it('should respect the timeout option passed', async () => {
       activateMock.mockReturnValue(null);
+      mockDelay = 20;
       readySuccess = false;
 
       render(
         <OptimizelyProvider optimizely={optimizelyMock}>
-          <MyExperimentComponent options={{ timeout: mockDelay }} />
+          <MyExperimentComponent options={{ timeout: mockDelay - 10 }} />
         </OptimizelyProvider>
       );
 
-      await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('null|false|false')); // initial render
-
-      await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('null|false|true')); // when didTimeout
-
-      // Simulate datafile fetch completing after timeout has already passed
-      // Activate now returns a variation
+      expect(screen.getByTestId('result')).toHaveTextContent('null|false|false'); // initial render
+      readySuccess = true;
       activateMock.mockReturnValue('12345');
-      await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('12345|true|true')); // when clientReady
+      // When timeout is reached, but dataReadyPromise is resolved later with the variation
+      await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('12345|true|true'));
     });
 
     it('should gracefully handle the client promise rejecting after timeout', async () => {
-      jest.useFakeTimers();
-
       readySuccess = false;
       activateMock.mockReturnValue('12345');
+
       getOnReadyPromise = (): Promise<void> =>
         new Promise((_, rej) => setTimeout(() => rej(REJECTION_REASON), mockDelay));
 
@@ -231,11 +256,7 @@ describe('hooks', () => {
         </OptimizelyProvider>
       );
 
-      jest.advanceTimersByTime(mockDelay + 1);
-
       await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('null|false|false'));
-
-      jest.useRealTimers();
     });
 
     it('should re-render when the user attributes change using autoUpdate', async () => {
@@ -247,9 +268,6 @@ describe('hooks', () => {
         </OptimizelyProvider>
       );
 
-      // TODO - Wrap this with async act() once we upgrade to React 16.9
-      // See https://github.com/facebook/react/issues/15379
-      await optimizelyMock.onReady();
       await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('null|true|false'));
 
       activateMock.mockReturnValue('12345');
@@ -257,8 +275,7 @@ describe('hooks', () => {
       await act(async () => {
         userUpdateCallbacks.forEach((fn) => fn());
       });
-      //   component.update();
-      //   await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('12345|true|false');
+
       await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('12345|true|false'));
     });
 
@@ -270,9 +287,6 @@ describe('hooks', () => {
         </OptimizelyProvider>
       );
 
-      //   // TODO - Wrap this with async act() once we upgrade to React 16.9
-      //   // See https://github.com/facebook/react/issues/15379
-      await optimizelyMock.onReady();
       await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('null|true|false'));
 
       activateMock.mockReturnValue('12345');
@@ -426,6 +440,16 @@ describe('hooks', () => {
   });
 
   describe('useFeature', () => {
+    it('should print error if optimizely is not provided', async () => {
+      render(
+        // @ts-ignore
+        <OptimizelyProvider optimizely={null}>
+          <MyFeatureComponent />
+        </OptimizelyProvider>
+      );
+      await waitFor(() => expect(hooksLogger.error).toHaveBeenCalled());
+    });
+
     it('should render true when the feature is enabled', async () => {
       isFeatureEnabledMock.mockReturnValue(true);
 
@@ -434,7 +458,6 @@ describe('hooks', () => {
           <MyFeatureComponent />
         </OptimizelyProvider>
       );
-      await optimizelyMock.onReady();
       await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('true|{"foo":"bar"}|true|false'));
     });
 
@@ -447,44 +470,33 @@ describe('hooks', () => {
           <MyFeatureComponent />
         </OptimizelyProvider>
       );
-      await optimizelyMock.onReady();
       await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('false|{}|true|false'));
     });
 
     it('should respect the timeout option passed', async () => {
+      mockDelay = 20;
       isFeatureEnabledMock.mockReturnValue(false);
       featureVariables = {};
       readySuccess = false;
 
       render(
         <OptimizelyProvider optimizely={optimizelyMock}>
-          <MyFeatureComponent options={{ timeout: mockDelay }} />
+          <MyFeatureComponent options={{ timeout: mockDelay - 10 }} />
         </OptimizelyProvider>
       );
 
-      await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('false|{}|false|false')); // initial render
+      // Initial render
+      expect(screen.getByTestId('result')).toHaveTextContent('false|{}|false|false');
 
-      await optimizelyMock.onReady();
-      await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('false|{}|false|true')); // when didTimeout
-
-      // Simulate datafile fetch completing after timeout has already passed
-      // isFeatureEnabled now returns true, getFeatureVariables returns variable values
+      readySuccess = true;
       isFeatureEnabledMock.mockReturnValue(true);
       featureVariables = mockFeatureVariables;
-      // Wait for completion of dataReadyPromise
-      await optimizelyMock.onReady().then((res) => res.dataReadyPromise);
 
-      // Simulate datafile fetch completing after timeout has already passed
-      // Activate now returns a variation
-      activateMock.mockReturnValue('12345');
-      // Wait for completion of dataReadyPromise
-      await optimizelyMock.onReady().then((res) => res.dataReadyPromise);
-      await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('true|{"foo":"bar"}|true|true')); // when clientReady
+      // When timeout is reached, but dataReadyPromise is resolved later with the feature enabled
+      await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('true|{"foo":"bar"}|true|true'));
     });
 
     it('should gracefully handle the client promise rejecting after timeout', async () => {
-      jest.useFakeTimers();
-
       readySuccess = false;
       isFeatureEnabledMock.mockReturnValue(true);
       getOnReadyPromise = (): Promise<void> =>
@@ -496,11 +508,7 @@ describe('hooks', () => {
         </OptimizelyProvider>
       );
 
-      jest.advanceTimersByTime(mockDelay + 1);
-
       await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('false|{}|false|false'));
-
-      jest.useRealTimers();
     });
 
     it('should re-render when the user attributes change using autoUpdate', async () => {
@@ -513,9 +521,6 @@ describe('hooks', () => {
         </OptimizelyProvider>
       );
 
-      // TODO - Wrap this with async act() once we upgrade to React 16.9
-      // See https://github.com/facebook/react/issues/15379
-      await optimizelyMock.onReady();
       await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('false|{}|true|false'));
 
       isFeatureEnabledMock.mockReturnValue(true);
@@ -537,9 +542,6 @@ describe('hooks', () => {
         </OptimizelyProvider>
       );
 
-      // TODO - Wrap this with async act() once we upgrade to React 16.9
-      // See https://github.com/facebook/react/issues/15379
-      await optimizelyMock.onReady();
       await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('false|{}|true|false'));
 
       isFeatureEnabledMock.mockReturnValue(true);
@@ -548,7 +550,7 @@ describe('hooks', () => {
       act(() => {
         userUpdateCallbacks.forEach((fn) => fn());
       });
-      // component.update();
+
       await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('false|{}|true|false'));
     });
 
@@ -677,6 +679,160 @@ describe('hooks', () => {
   });
 
   describe('useDecision', () => {
+    it('should handle no client promise response', async () => {
+      getOnReadyPromise = () =>
+        new Promise((resolve) => {
+          setTimeout(() => {
+            resolve({
+              success: false,
+              reason: NotReadyReason.NO_CLIENT,
+              dataReadyPromise: new Promise((r) => setTimeout(() => r({ success: false }), mockDelay)),
+            });
+          });
+        });
+      decideMock.mockReturnValue({ ...defaultDecision });
+
+      render(
+        <OptimizelyProvider optimizely={optimizelyMock}>
+          <MyDecideComponent />
+        </OptimizelyProvider>
+      );
+
+      await waitFor(() => {
+        expect(screen.getByTestId('result')).toHaveTextContent('false|{}|false|false');
+        expect(hooksLogger.warn).toHaveBeenCalled();
+      });
+    });
+
+    it('should handle no client, but data ready promise success', () => {
+      getOnReadyPromise = () =>
+        new Promise((resolve) => {
+          setTimeout(() => {
+            resolve({
+              success: false,
+              reason: NotReadyReason.NO_CLIENT,
+              dataReadyPromise: new Promise((r) => setTimeout(() => r({ success: true }), mockDelay)),
+            });
+          });
+        });
+      decideMock.mockReturnValue({ ...defaultDecision });
+
+      render(
+        <OptimizelyProvider optimizely={optimizelyMock}>
+          <MyDecideComponent />
+        </OptimizelyProvider>
+      );
+
+      expect(screen.getByTestId('result')).toHaveTextContent('false|{}|true|false');
+    });
+
+    it('should handle user not ready promise response', async () => {
+      getOnReadyPromise = () =>
+        new Promise((resolve) => {
+          setTimeout(() => {
+            resolve({
+              success: false,
+              reason: NotReadyReason.USER_NOT_READY,
+              dataReadyPromise: new Promise((r) => setTimeout(() => r({ success: false }), mockDelay)),
+            });
+          });
+        });
+      decideMock.mockReturnValue({ ...defaultDecision });
+
+      render(
+        <OptimizelyProvider optimizely={optimizelyMock}>
+          <MyDecideComponent />
+        </OptimizelyProvider>
+      );
+
+      await waitFor(() => {
+        expect(screen.getByTestId('result')).toHaveTextContent('false|{}|false|false');
+        expect(hooksLogger.warn).toHaveBeenCalled();
+      });
+    });
+
+    it('should handle user not ready, but data ready promise success', () => {
+      getOnReadyPromise = () =>
+        new Promise((resolve) => {
+          setTimeout(() => {
+            resolve({
+              success: false,
+              reason: NotReadyReason.USER_NOT_READY,
+              dataReadyPromise: new Promise((r) => setTimeout(() => r({ success: true }), mockDelay)),
+            });
+          });
+        });
+      decideMock.mockReturnValue({ ...defaultDecision });
+
+      render(
+        <OptimizelyProvider optimizely={optimizelyMock}>
+          <MyDecideComponent />
+        </OptimizelyProvider>
+      );
+
+      expect(screen.getByTestId('result')).toHaveTextContent('false|{}|true|false');
+    });
+
+    it('should handle default success false case', async () => {
+      getOnReadyPromise = () =>
+        new Promise((resolve) => {
+          setTimeout(() => {
+            resolve({
+              success: false,
+              reason: 'UNKNOWN',
+              dataReadyPromise: new Promise((r) => setTimeout(() => r({ success: false }), mockDelay)),
+            });
+          });
+        });
+      decideMock.mockReturnValue({ ...defaultDecision });
+
+      render(
+        <OptimizelyProvider optimizely={optimizelyMock}>
+          <MyDecideComponent />
+        </OptimizelyProvider>
+      );
+
+      await waitFor(() => {
+        expect(hooksLogger.warn).toHaveBeenCalled();
+        expect(screen.getByTestId('result')).toHaveTextContent('false|{}|false|false');
+      });
+    });
+
+    it('should handle default success true case', async () => {
+      getOnReadyPromise = () =>
+        new Promise((resolve) => {
+          setTimeout(() => {
+            resolve({
+              success: false,
+              reason: 'UNKNOWN',
+              dataReadyPromise: new Promise((r) => setTimeout(() => r({ success: true }), mockDelay)),
+            });
+          });
+        });
+      decideMock.mockReturnValue({ ...defaultDecision });
+
+      render(
+        <OptimizelyProvider optimizely={optimizelyMock}>
+          <MyDecideComponent />
+        </OptimizelyProvider>
+      );
+
+      await waitFor(() => {
+        expect(hooksLogger.warn).toHaveBeenCalled();
+        expect(screen.getByTestId('result')).toHaveTextContent('false|{}|true|false');
+      });
+    });
+
+    it('should print error if optimizely is not provided', async () => {
+      render(
+        // @ts-ignore
+        <OptimizelyProvider optimizely={null}>
+          <MyDecideComponent />
+        </OptimizelyProvider>
+      );
+      await waitFor(() => expect(hooksLogger.error).toHaveBeenCalled());
+    });
+
     it('should render true when the flag is enabled', async () => {
       decideMock.mockReturnValue({
         ...defaultDecision,
@@ -688,8 +844,6 @@ describe('hooks', () => {
           <MyDecideComponent />
         </OptimizelyProvider>
       );
-      await optimizelyMock.onReady();
-      // component.update();
       await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('true|{"foo":"bar"}|true|false'));
     });
 
@@ -699,50 +853,65 @@ describe('hooks', () => {
         enabled: false,
         variables: { foo: 'bar' },
       });
+
       render(
         <OptimizelyProvider optimizely={optimizelyMock}>
           <MyDecideComponent />
         </OptimizelyProvider>
       );
-      await optimizelyMock.onReady();
+
       await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('false|{"foo":"bar"}|true|false'));
     });
 
     it('should respect the timeout option passed', async () => {
       decideMock.mockReturnValue({ ...defaultDecision });
       readySuccess = false;
+      mockDelay = 20;
 
       render(
         <OptimizelyProvider optimizely={optimizelyMock}>
-          <MyDecideComponent options={{ timeout: mockDelay }} />
+          <MyDecideComponent options={{ timeout: mockDelay - 10 }} />
         </OptimizelyProvider>
       );
-      await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('false|{}|false|false'));
 
-      await optimizelyMock.onReady();
-      // component.update();
-      await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('false|{}|false|true'));
+      // Initial render
+      expect(screen.getByTestId('result')).toHaveTextContent('false|{}|false|false');
 
-      // Simulate datafile fetch completing after timeout has already passed
-      // flag is now true and decision contains variables
       decideMock.mockReturnValue({
         ...defaultDecision,
         enabled: true,
         variables: { foo: 'bar' },
       });
+      readySuccess = true;
 
-      await optimizelyMock.onReady().then((res) => res.dataReadyPromise);
+      // When timeout is reached, but dataReadyPromise is resolved later with the decision value
+      await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('true|{"foo":"bar"}|true|true'));
+    });
 
-      // Simulate datafile fetch completing after timeout has already passed
-      // Wait for completion of dataReadyPromise
-      await optimizelyMock.onReady().then((res) => res.dataReadyPromise);
+    it('should log warn message if dataReadyPromise resolves as false', async () => {
+      decideMock.mockReturnValue({ ...defaultDecision });
+      readySuccess = false;
+      mockDelay = 20;
 
-      await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('true|{"foo":"bar"}|true|true')); // when clientReady
+      render(
+        <OptimizelyProvider optimizely={optimizelyMock}>
+          <MyDecideComponent options={{ timeout: mockDelay - 10 }} />
+        </OptimizelyProvider>
+      );
+
+      // Initial render
+      expect(screen.getByTestId('result')).toHaveTextContent('false|{}|false|false');
+
+      readySuccess = false;
+
+      // When timeout is reached, but dataReadyPromise is resolved later with the decision value
+      await waitFor(() => {
+        expect(hooksLogger.warn).toHaveBeenCalled();
+        expect(screen.getByTestId('result')).toHaveTextContent('false|{}|false|true');
+      });
     });
 
     it('should gracefully handle the client promise rejecting after timeout', async () => {
-      jest.useFakeTimers();
-
       readySuccess = false;
       decideMock.mockReturnValue({ ...defaultDecision });
       getOnReadyPromise = (): Promise<void> =>
@@ -754,11 +923,9 @@ describe('hooks', () => {
         </OptimizelyProvider>
       );
 
-      jest.advanceTimersByTime(mockDelay + 1);
-
-      await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('false|{}|false|false'));
-
-      jest.useRealTimers();
+      await waitFor(() => {
+        expect(screen.getByTestId('result')).toHaveTextContent('false|{}|false|false');
+      });
     });
 
     it('should re-render when the user attributes change using autoUpdate', async () => {
@@ -769,9 +936,6 @@ describe('hooks', () => {
         </OptimizelyProvider>
       );
 
-      // TODO - Wrap this with async act() once we upgrade to React 16.9
-      // See https://github.com/facebook/react/issues/15379
-      await optimizelyMock.onReady();
       await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('false|{}|true|false'));
 
       decideMock.mockReturnValue({
@@ -794,9 +958,6 @@ describe('hooks', () => {
         </OptimizelyProvider>
       );
 
-      // TODO - Wrap this with async act() once we upgrade to React 16.9
-      // See https://github.com/facebook/react/issues/15379
-      await optimizelyMock.onReady();
       await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('false|{}|true|false'));
 
       decideMock.mockReturnValue({
@@ -819,7 +980,7 @@ describe('hooks', () => {
           <UseDecisionLoggingComponent />
         </OptimizelyProvider>
       );
-      // component.update();
+
       expect(mockLog).toHaveBeenCalledTimes(1);
       expect(mockLog).toHaveBeenCalledWith(false);
     });
@@ -990,7 +1151,6 @@ describe('hooks', () => {
       );
 
       decideMock.mockReturnValue({ ...defaultDecision, variables: { foo: 'bar' } });
-      await optimizelyMock.onReady();
       await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('false|{"foo":"bar"}|true|false'));
     });
 
@@ -1016,7 +1176,6 @@ describe('hooks', () => {
         { variationKey: 'var2' }
       );
 
-      await optimizelyMock.onReady();
       await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('false|{}|true|false'));
     });
 
@@ -1044,20 +1203,20 @@ describe('hooks', () => {
         { variationKey: 'var2' }
       );
 
-      await optimizelyMock.onReady();
       await waitFor(() => expect(screen.getByTestId('result')).toHaveTextContent('false|{}|true|false'));
     });
   });
+
   describe('useTrackEvent', () => {
     it('returns track method and false states when optimizely is not provided', () => {
       const wrapper = ({ children }: { children: React.ReactNode }): React.ReactElement => {
         //@ts-ignore
         return <OptimizelyProvider>{children}</OptimizelyProvider>;
       };
+
       const { result } = renderHook(() => useTrackEvent(), { wrapper });
-      expect(result.current[0]).toBeInstanceOf(Function);
-      expect(result.current[1]).toBe(false);
-      expect(result.current[2]).toBe(false);
+
+      expect(result.current).toEqual([expect.any(Function), false, false]);
     });
 
     it('returns track method along with clientReady and didTimeout state when optimizely instance is provided', () => {
@@ -1068,9 +1227,10 @@ describe('hooks', () => {
       );
 
       const { result } = renderHook(() => useTrackEvent(), { wrapper });
-      expect(result.current[0]).toBeInstanceOf(Function);
-      expect(typeof result.current[1]).toBe('boolean');
-      expect(typeof result.current[2]).toBe('boolean');
+      result.current[0]('eventKey');
+
+      expect(optimizelyMock.track).toHaveBeenCalledTimes(1);
+      expect(result.current).toEqual([expect.any(Function), true, false]);
     });
 
     it('Log error when track method is called and optimizely instance is not provided', () => {
@@ -1080,7 +1240,7 @@ describe('hooks', () => {
       };
       const { result } = renderHook(() => useTrackEvent(), { wrapper });
       result.current[0]('eventKey');
-      expect(hooksLogger.error).toHaveBeenCalledTimes(1);
+      expect(hooksLogger.error).toHaveBeenCalled();
     });
 
     it('Log error when track method is called and client is not ready', () => {
@@ -1094,7 +1254,8 @@ describe('hooks', () => {
 
       const { result } = renderHook(() => useTrackEvent(), { wrapper });
       result.current[0]('eventKey');
-      expect(hooksLogger.error).toHaveBeenCalledTimes(1);
+
+      expect(hooksLogger.error).toHaveBeenCalled();
     });
   });
 });