+++ src/utils.tsx
@@ -32,52 +32,22 @@ export function areUsersEqual(user1: UserInfo, user2: UserInfo): boolean {
     return false;
   }
 
-  const user1keys = Object.keys(user1.attributes || {});
-  const user2keys = Object.keys(user2.attributes || {});
-  user1keys.sort();
-  user2keys.sort();
-
   const user1Attributes = user1.attributes || {};
   const user2Attributes = user2.attributes || {};
 
-  const areKeysLenEqual = user1keys.length === user2keys.length;
-  if (!areKeysLenEqual) {
+  const user1Keys = Object.keys(user1Attributes);
+  const user2Keys = Object.keys(user2Attributes);
+
+  if (user1Keys.length !== user2Keys.length) {
     return false;
   }
 
-  for (let i = 0; i < user1keys.length; i++) {
-    const key1 = user1keys[i];
-    const key2 = user2keys[i];
-    if (key1 !== key2) {
-      return false;
-    }
-
-    if (user1Attributes[key1] !== user2Attributes[key2]) {
+  for (const key of user1Keys) {
+    if (user1Attributes[key] !== user2Attributes[key]) {
       return false;
     }
   }
 
-  return true;
-}
-
-export function areObjectsEqual(obj1: any, obj2: any) {
-  const obj1Keys = Object.keys(obj1);
-  const obj2Keys = Object.keys(obj2);
-
-  if (obj1Keys.length !== obj2Keys.length) {
-    return false;
-  }
-
-  for (let i = 0; i < obj1Keys.length; i += 1) {
-    const key = obj1Keys[i];
-    if (typeof obj1[key] === 'object' && typeof obj2[key] === 'object') {
-      if (!areObjectsEqual(obj1[key], obj2[key])) {
-        return false;
-      }
-    } else if (obj1[key] !== obj2[key]) {
-      return false;
-    }
-  }
   return true;
 }